DynamicProjectSchedulingApp/requirements.txt
asgiref==3.2.9
dj-database-url==0.5.0
Django==3.0.8
djangorestframework==3.11.0
Faker==4.1.1
flake8==3.8.3
gunicorn==20.0.4
json-flatten==0.1
mccabe==0.6.1
Pillow==7.2.0
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
whitenoise==5.1.0
